Move script files to src/ folder
