+++ index.html
@@ -5,7 +5,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Sisukas</title>
     <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap-icons/font/bootstrap-icons.css">
-    <link rel="stylesheet" href="styles.css?v=0.4">
+    <link rel="stylesheet" href="styles.css?v=0.5">
 </head>
 <body>
     <div id="title-container">
@@ -73,7 +73,7 @@
 
     </div>
 
-    <script type="module" src="script.js?v=0.4"></script>
+    <script type="module" src="src/script.js?v=0.5"></script>
 
 </body>
 </html>
