+++ src/loader.js
